backend/requirements.txt
# Flask and extensions
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Talisman==1.1.0

# RSS feed parsing
feedparser==6.0.11
aiohttp==3.11.10
python-dateutil==2.9.0

# Security and validation
bleach==6.2.0
python-dotenv==1.0.1

# WSGI server
gunicorn==23.0.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0

# Development
black==24.10.0
flake8==7.1.1
